agent: refactor integration tests to include fv tests
