+++ calico-vpp-agent/policy/policy_server.go
@@ -166,11 +166,6 @@ func NewPolicyServer(vpp *vpplink.VppLink, log *logrus.Entry) (*Server, error) {
 		return nil, errors.Wrapf(err, "Could not delete socket %s", config.FelixDataplaneSocket)
 	}
 
-	err = InstallFelixPlugin()
-	if err != nil {
-		return nil, errors.Wrap(err, "could not install felix plugin")
-	}
-
 	return server, nil
 }
