+++ calico-vpp-agent/watch_dog/watch_dog.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package watchdog
 
 import (
 	"time"
